feat: move the endpoint's param table into render-core
